util/tests/vma: Fix warning c++11-narrowing
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=106801
Fixes: 943fecc569 ("util: Add a randomized test for the virtual memory allocator")

+++ src/util/tests/vma/vma_random_test.cpp
@@ -236,7 +236,7 @@ int main(int argc, char **argv)
       errx(1, "USAGE: %s seed iter_count\n", argv[0]);
    }
 
-   random_test r{seed};
+   random_test r{(uint_fast32_t)seed};
    r.test(count);
 
    printf("ok\n");
